+++ src/content/input-row/content-readonly-row.component.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import InputRow from './content-input-row.component';
 
 const ContentReadonlyRow = ({
-  label, value, normalize, asColumn,
+  label, value, normalize, asColumn, id,
 }) => (
-  <InputRow asColumn={asColumn} label={label} showError={false}>
+  <InputRow id={id} asColumn={asColumn} label={label} showError={false}>
     <span>
       {normalize ? normalize(value) : value}
     </span>
@@ -17,6 +17,7 @@ ContentReadonlyRow.propTypes = {
   label: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
   value: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.node, PropTypes.bool]),
   normalize: PropTypes.func,
+  id: PropTypes.string.isRequired,
 };
 
 ContentReadonlyRow.defaultProps = {